requirements.txt
arrow==1.2.0
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
black==21.9b0
blessed==1.19.0
Brotli==1.0.9
cffi==1.14.6
cfgv==3.3.1
click==8.0.3
cssselect2==0.4.1
distlib==0.3.3
Django==3.2.8
django-picklefield==3.0.1
django-q==1.3.9
factory-boy==3.2.1
Faker==9.8.0
filelock==3.3.1
fonttools==4.27.1
html5lib==1.1
identify==2.3.5
install==1.3.4
mypy-extensions==0.4.3
nodeenv==1.6.0
pathspec==0.9.0
Pillow==8.3.2
platformdirs==2.4.0
pre-commit==2.15.0
pycparser==2.20
pydyf==0.1.1
pyphen==0.11.0
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
pywatchman==1.4.1
PyYAML==6.0
redis==3.5.3
regex==2021.10.23
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
tinycss2==1.1.0
toml==0.10.2
tomli==1.2.1
tornado==6.1
typing-extensions==3.10.0.2
virtualenv==20.8.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
zopfli==0.1.8
